src/cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blueplane-cli"
version = "0.1.0"
description = "Privacy-first telemetry CLI for AI-assisted coding"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "AGPL-3.0-only" }
authors = [
    { name = "Sierra Labs LLC" }
]
keywords = ["telemetry", "analytics", "cli", "ai", "coding", "productivity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
charts = [
    "plotext>=5.0.0",
]
websocket = [
    "websockets>=11.0",
]
export = [
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",  # For parquet support
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "blueplane-cli[charts,websocket,export]",
]

[project.scripts]
blueplane = "cli.main:main"
bp = "cli.main:main"  # Short alias

[project.urls]
Homepage = "https://github.com/sierra-labs/blueplane-telemetry-core"
Documentation = "https://github.com/sierra-labs/blueplane-telemetry-core/blob/main/docs/architecture/layer3_cli_interface.md"
Repository = "https://github.com/sierra-labs/blueplane-telemetry-core"
Issues = "https://github.com/sierra-labs/blueplane-telemetry-core/issues"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line length handled by black
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=cli --cov-report=term-missing"

[tool.coverage.run]
source = ["cli"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]